improved test. functionality tests all pass. Example tests moved to own directory
