+++ tests/runExamples.js
@@ -1,0 +1,2 @@
+const { runTestDir } = require("../index");
+runTestDir('tests/exampleTests');
