+++ tests/testingTheTests/compatibility/1.3/assertions/equality-assertions.test.ts
@@ -1,4 +1,4 @@
-import { expect, test } from '../../../index';
+import { expect, test } from '../../../../../index';
 
 test("expect.equal: Primitive", function() {
     expect.equal(5,5);
@@ -24,23 +24,32 @@ test("expect.equal: Reference", function() {
 })
 
 
-test("(Fail Case) expect.equal: Reference", function() {
+test("(Fail Case) expect.equal", function() {
     let fail = new Map()
     let fail2 = new Map()
-    expect.equal(fail,fail2);
+
+    expect.throwsError(()=>{
+        expect.equal(fail,fail2);
+    });
 })
 
 
 test("(Fail Case) expect.equal: Array", function() {
     let arr = [1,2,3,4,5];
     let arr2 = [...arr];
-    expect.equal(arr2,arr);
+    
+    expect.throwsError(()=>{
+        expect.equal(arr2,arr);
+    });
 })
 
 test("(Fail Case) expect.equal: ArrayString", function() {
     let arr = [1,2,3,4,5];
     let str = "1,2,3,4,5";
-    expect.equal(str,arr);
+    
+    expect.throwsError(()=>{
+        expect.equal(str,arr);
+    });
 })
 
 test("expect.equalLoose", function() {
