+++ tests/testingTheTests/compatibility/1.3/assertions/nullity-assertions.test.ts
@@ -1,10 +1,12 @@
-import { expect, test } from '../../../index';
+import { expect, test } from '../../../../../index';
 
 test("expect.null", function() {
     expect.null(null);
 })
-test("(Fail Case) expect.null", function() {
-    expect.null(5);
+test("(Fail Case) expect.null", function() {   
+    expect.throwsError(()=>{
+        expect.null(5);
+    });
 })
 
 test("expect.notNull", function() {
@@ -12,14 +14,19 @@ test("expect.notNull", function() {
     expect.notNull(undefined);
 })
 test("(Fail Case) expect.notNull", function() {
-    expect.notNull(null);
+    expect.throwsError(()=>{
+        expect.notNull(null);
+    });
+
 })
 
 test("expect.undefined", function() {
     expect.undefined(undefined);
 })
 test("(Fail Case) expect.undefined", function() {
-    expect.undefined(null);
+    expect.throwsError(()=>{
+        expect.undefined(null);
+    });
 })
 
 test("expect.notUndefined", function() {
@@ -27,5 +34,7 @@ test("expect.notUndefined", function() {
     expect.notUndefined(null);
 })
 test("(Fail Case) expect.notUndefined", function() {
-    expect.notUndefined(undefined);
+    expect.throwsError(()=>{
+        expect.notUndefined(undefined);
+    });
 })
